PEP#: 233 TITLE: Updated Embedded Instance Support branch
DESCRIPTION:

- Merged with latest code from 2.5.1

- Modified ObjectNormalizer to use a CIMOMHandle instead of the repository and to normalize the instances in an embedded instance property.

+++ src/Pegasus/Common/TimeValue.h
@@ -1,31 +1,37 @@
-//%LICENSE////////////////////////////////////////////////////////////////
+//%2006////////////////////////////////////////////////////////////////////////
 //
-// Licensed to The Open Group (TOG) under one or more contributor license
-// agreements.  Refer to the OpenPegasusNOTICE.txt file distributed with
-// this work for additional information regarding copyright ownership.
-// Each contributor licenses this file to you under the OpenPegasus Open
-// Source License; you may not use this file except in compliance with the
-// License.
+// Copyright (c) 2000, 2001, 2002 BMC Software; Hewlett-Packard Development
+// Company, L.P.; IBM Corp.; The Open Group; Tivoli Systems.
+// Copyright (c) 2003 BMC Software; Hewlett-Packard Development Company, L.P.;
+// IBM Corp.; EMC Corporation, The Open Group.
+// Copyright (c) 2004 BMC Software; Hewlett-Packard Development Company, L.P.;
+// IBM Corp.; EMC Corporation; VERITAS Software Corporation; The Open Group.
+// Copyright (c) 2005 Hewlett-Packard Development Company, L.P.; IBM Corp.;
+// EMC Corporation; VERITAS Software Corporation; The Open Group.
+// Copyright (c) 2006 Hewlett-Packard Development Company, L.P.; IBM Corp.;
+// EMC Corporation; Symantec Corporation; The Open Group.
 //
-// Permission is hereby granted, free of charge, to any person obtaining a
-// copy of this software and associated documentation files (the "Software"),
-// to deal in the Software without restriction, including without limitation
-// the rights to use, copy, modify, merge, publish, distribute, sublicense,
-// and/or sell copies of the Software, and to permit persons to whom the
-// Software is furnished to do so, subject to the following conditions:
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to
+// deal in the Software without restriction, including without limitation the
+// rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+// sell copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+// 
+// THE ABOVE COPYRIGHT NOTICE AND THIS PERMISSION NOTICE SHALL BE INCLUDED IN
+// ALL COPIES OR SUBSTANTIAL PORTIONS OF THE SOFTWARE. THE SOFTWARE IS PROVIDED
+// "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
+// LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+// PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+// ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-// The above copyright notice and this permission notice shall be included
-// in all copies or substantial portions of the Software.
+//==============================================================================
 //
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-// CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-// SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+// Author: Mike Brasher (mbrasher@bmc.com)
 //
-//////////////////////////////////////////////////////////////////////////
+// Modified By:
 //
 //%/////////////////////////////////////////////////////////////////////////////
 
@@ -48,54 +54,53 @@ public:
     {
     }
 
-    TimeValue(Uint32 seconds, Uint32 microseconds)
-        : _seconds(seconds), _microseconds(microseconds)
+    TimeValue(Uint32 seconds, Uint32 microseconds) 
+	: _seconds(seconds), _microseconds(microseconds)
     {
     }
 
-    Uint32 getSeconds() const
+    Uint32 getSeconds() const 
     {
-        return _seconds;
+	return _seconds;
     }
 
     void setSeconds(Uint32 seconds)
     {
-        _seconds = seconds;
+	_seconds = seconds;
     }
 
-    Uint32 getMicroseconds() const
+    Uint32 getMicroseconds() const 
     {
-        return _microseconds;
+	return _microseconds;
     }
 
-    Uint64 toMicroseconds() const
+    Uint64 toMicroseconds() const 
     {
-        return (Uint64(_seconds) * Uint64(1000000)) + Uint64(_microseconds);
+	return (Uint64(_seconds) * Uint64(1000000)) + Uint64(_microseconds);
     }
 
     void setMicroseconds(Uint32 microseconds)
     {
-        _microseconds = microseconds;
+	_microseconds = microseconds;
     }
 
-    void fromMilliseconds(Uint64 milliseconds)
+    void fromMilliseconds(Uint32 milliseconds)
     {
-        _seconds = Uint32(milliseconds / 1000);
-        _microseconds = Uint32(milliseconds % 1000) * 1000;
+	_seconds = milliseconds / 1000;
+	_microseconds = (milliseconds % 1000) * 1000;
     }
 
-    Uint64 toMilliseconds() const
+    Uint32 toMilliseconds() const 
     {
-        return Uint64(_seconds) * Uint64(1000) +
-            Uint64(_microseconds) / Uint64(1000);
+	return _seconds * 1000 + _microseconds / 1000; 
     }
 
     static TimeValue getCurrentTime()
     {
-        Uint32 seconds;
-        Uint32 microseconds;
-        System::getCurrentTimeUsec(seconds, microseconds);
-        return TimeValue(seconds, microseconds);
+	Uint32 seconds;
+	Uint32 microseconds;
+	System::getCurrentTimeUsec(seconds, microseconds);
+	return TimeValue(seconds, microseconds);
     }
 
 private:
